CONFETI-28 refactor: Event 패키지 구조 수정 - 도메인 하위에서 messagebroker 하위로 변경. dto 보다는 이벤트의 성격이 더 강하므로
