+++ src/main/java/org/sopt/confeti/global/messagebroker/sqs/event/artist/CreateArtistEvent.java
@@ -1,25 +1,24 @@
-package org.sopt.confeti.domain.applemusic.song.event;
+package org.sopt.confeti.global.messagebroker.sqs.event.artist;
 
 
-import org.sopt.confeti.domain.applemusic.song.Song;
+import org.sopt.confeti.domain.applemusic.artist.Artist;
 import org.sopt.confeti.global.messagebroker.message.CreateEvent;
 
-public record CreateSongEvent(
-    String songId,
+public record CreateArtistEvent(
+    String artistId,
     String name,
     String artworkUrl,
     Integer artworkWidth,
     Integer artworkHeight
 ) implements CreateEvent {
 
-    public Song toSong() {
-        return Song.builder()
-            .songId(songId)
+    public Artist toArtist() {
+        return Artist.builder()
+            .artistId(artistId)
             .name(name)
             .artworkUrl(artworkUrl)
             .artworkWidth(artworkWidth)
             .artworkHeight(artworkHeight)
             .build();
     }
-
 }
