+++ src/main/java/org/sopt/confeti/global/messagebroker/sqs/handler/artist/CreateArtistEventHandler.java
@@ -3,8 +3,8 @@ package org.sopt.confeti.global.messagebroker.sqs.handler.artist;
 
 import lombok.RequiredArgsConstructor;
 import org.sopt.confeti.domain.applemusic.artist.application.ArtistService;
-import org.sopt.confeti.domain.applemusic.artist.event.CreateArtistEvent;
 import org.sopt.confeti.global.messagebroker.handler.CreateEventHandler;
+import org.sopt.confeti.global.messagebroker.sqs.event.artist.CreateArtistEvent;
 import org.sopt.confeti.global.messagebroker.sqs.handler.SqsEventHandler;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
