+++ src/main/java/org/sopt/confeti/global/messagebroker/sqs/handler/song/CreateSongEventHandler.java
@@ -2,8 +2,8 @@ package org.sopt.confeti.global.messagebroker.sqs.handler.song;
 
 import lombok.RequiredArgsConstructor;
 import org.sopt.confeti.domain.applemusic.song.application.SongService;
-import org.sopt.confeti.domain.applemusic.song.event.CreateSongEvent;
 import org.sopt.confeti.global.messagebroker.handler.CreateEventHandler;
+import org.sopt.confeti.global.messagebroker.sqs.event.song.CreateSongEvent;
 import org.sopt.confeti.global.messagebroker.sqs.handler.SqsEventHandler;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
